requirements.txt
selenium==3.141.0
Edge